Install task works

+++ lib/merb_pupu/github.rb
@@ -2,23 +2,41 @@ require "fileutils"
 require "yaml"
 require "ostruct"
 
+module ShellExtensions
+  def run(command)
+    puts "[bash] #{command}"
+    %x(#{command})
+  end
+end
+
 module Merb
   module Plugins
     class GitHub
       class << self
+        include ShellExtensions
         # GitHub.install("autocompleter")
         # GitHub.install("botanicus/autocompleter")
         def install(repo)
           user, repo = repo.split("/") if repo.match(%r{/})
-          user = ENV["USER"] unless defined?(user)
+          user = ENV["USER"] unless user
           url = "git://github.com/#{user}/pupu-#{repo}.git"
-          chdir do |public_dir, repo_dir|
-            puts %x(git clone #{url})
-            Dir.chdir(repo_dir) do
-              FileUtils.mv "initializers/#{repo}.js",  "#{public_dir}/javascripts/#{repo}.js"
-              FileUtils.mv "initializers/#{repo}.css", "#{public_dir}/stylesheets/#{repo}.css"
-              revision = "git log | head -1".chomp.sub(/^commit /, "")
+          chdir do |public_dir|
+            raise "PluginIsAlreadyInstalled" if File.directory?(repo) # TODO: custom exception class
+            run "git clone #{url} #{repo}"
+            Dir.chdir(repo) do
+              js_initializer = "initializers/#{repo}.js"
+              css_initializer = "initializers/#{repo}.css"
+              if File.exist?(js_initializer)
+                FileUtils.mkdir_p("#{public_dir}/javascripts")
+                FileUtils.mv js_initializer,  "#{public_dir}/javascripts/#{repo}.js"
+              end
+              if File.exist?(css_initializer)
+                FileUtils.mkdir_p("#{public_dir}/stylesheets")
+                FileUtils.mv css_initializer,  "#{public_dir}/stylesheets/#{repo}.css"
+              end
+              revision = %x(git log | head -1).chomp.sub(/^commit /, "")
               dependencies = Array.new # TODO
+              @pupu = Pupu[repo]
               self.save_metadata(:revision => revision, :repozitory => url, :dependencies => dependencies)
               FileUtils.rm_r ".git"
             end
@@ -43,7 +61,7 @@ module Merb
 
         protected
         def save_metadata(params)
-          Dir.chdir(@plugin.root) do
+          Dir.chdir(@pupu.root) do
             File.open("metadata.yml", "w") do |file|
               file.puts(params.to_yaml)
             end
@@ -52,8 +70,14 @@ module Merb
 
         def chdir(&block)
           public_dir = File.join(Dir.pwd, "public")
+          raise "PublicDirNotExists" unless File.directory?(public_dir) # TODO: create example class
+          FileUtils.mkdir_p(Pupu.root) unless File.directory?(Pupu.root)
           Dir.chdir(Pupu.root) do
-            block.call(public_dir, repo_dir)
+            if @pupu
+              block.call(public_dir, @pupu.root)
+            else
+              block.call(public_dir)
+            end
           end
         end
       end
